Add clang-tidy configuration
Fix reported warnings.

+++ include/util.hpp
@@ -7,7 +7,7 @@ namespace crypto::util {
 
 std::string read_base64_file(const std::string& name);
 
-int english_score(const std::vector<std::byte>& data);
+float english_score(const std::vector<std::byte>& data);
 
 int hamming_distance(const bytes_t& input1, const bytes_t& input2);
